requirements.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.11.0
Dijkstar==2.6.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.9
pygc==1.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tk==0.1.0
tqdm==4.66.1
